// frontend/src/types/user/user.d.ts
export interface User {
    id: number;
    username: string;
    first_name: string;
    last_name: string;
    major?: string;
    email: string;
    isActive: boolean;
    role: "student" | "admin";
    isSuperAdmin: boolean;
    isSuperuser: boolean;
    is_president?: boolean;
    is_vice_president?: boolean;
    is_event_manager?: boolean;
    president_of?: number;
    vice_president_of_society?: number;
    event_manager_of_society?: number;
    isStaff: boolean;
    following: number[];
    followers: number[];
    following_count?: number;
    followers_count?: number;
    fullName: string;
    isStudent(): boolean;
    isAdmin(): boolean;
    is_active: boolean;
    is_following?: boolean;
    icon?: string;
}